frontend/app/backend/requirements.txt
# Core Django
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1

# Environment variables
python-decouple==3.8

# Date/Time handling
python-dateutil==2.8.2
pytz==2024.1

# Validation
django-phonenumber-field==7.3.0
phonenumbers==8.13.27

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Code quality
flake8==7.0.0
black==23.12.1

# Production
gunicorn==21.2.0
whitenoise==6.6.0